Deduplicate `J3DMtxCalc`
